+++ CHANGES.md
@@ -6,6 +6,7 @@
 * Add ability to auto focus Search Show box on Home page and control this option via General Config/Interface
 * Add TVRage network name standardization
 * Change some provider images. Add a few new images
+* Change updated plex notifier (ported from SickBeard)
 
 [develop changelog]